Clarifies CodeGen sbt plugin add/enable instructions in client and schema docs
The `caliban-codegen-sbt` plugin needs to be added via the `project/plugins.sbt`
file and enabled through the `build.sbt` file in order to be used in an sbt
project. This change clarifies that the call to `addSbtPlugin` should be in
`project/plugins.sbt`, and the call to `enablePlugins` should be in `build.sbt`.