Upgrade to ZIO 2.0.0-RC6
The only way to create a `ZIO` is via the methods on the `ZIO` object.

`Clock`, `Console` and `Random` have been removed from method signatures
as they are now always available.

`ZManaged` has been replaced by `ZIO` with `Scope` in the environment.

`assertM` has been replaced with an assert after mapping on the `ZIO`.