Try medium+ resource class
The CI job is being killed, and it seems memory usage is maxing out.
I've checked a couple of runs of the failing step scripted3_jdk
on other branches, and they go to 98% memory usage.
Maybe something about this change causes more memory to be used.

The medium+ resource class is included on free plans, but it burns
15 credits a minute over 10/minute for the medium class.

I hope this doesn't incur unwarranted cost.