codegen: Add `EffectStrategy`
The scenario here is code generation of server code from graphql file. Context and proposal in #1925

If you generate code for server usage before this change, you'll need to evaluate and fetch all data to build a given response case class for a query.

However, Caliban supports suspending values of some fields in an effect in order to only evaluate those values if the client asked for the given field.

`EffectStrategy` allows you to plug in a mechanism to decide which fields should be wrapped in such a side-effect.